change routes and add trailing slashes
