+++ app/views/__init__.py
@@ -3,6 +3,7 @@ from app import app
 from .admin import *
 from .system import *
 from .user import *
+from .subscription import *
 
 
 @app.get("/", status_code=204)
